backend/requirements.txt
about-time
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
alive-progress
altgraph
annotated-types
anyio
async-timeout
attrs
beautifulsoup4
black
blinker
Brotli
budoux
certifi
charset-normalizer
click
colorama
colorlog
distro
EbookLib
et_xmlfile
exceptiongroup
fasttext-predict
filelock
Flask
Flask-Cors
frozenlist
fsspec
graphemeu
h11
httpcore
httpx
huggingface-hub
idna
importlib_metadata
importlib_resources
iniconfig
inquirerpy
itsdangerous
Jinja2
jiter
jsonschema
jsonschema-specifications
langid
litellm
lxml
markdown-it-py
MarkupSafe
mdurl
multidict
mypy-extensions
numpy
openai[aiohttp]
OpenCC
openpyxl
orjson
packaging
pathspec
pefile
pfzy
platformdirs
playsound3
pluggy
prompt_toolkit
propcache
pydantic
pydantic_core
Pygments
pyinstaller
pyinstaller-hooks-contrib
pytest
python-dotenv
pywin32-ctypes
PyYAML
re_edge_gpt
referencing
regex
requests
rich
rpds-py
six
sniffio
soupsieve
tenacity
tiktoken
tokenizers
tomli
tqdm
typing_extensions
urllib3
vaporetto
waitress
wcwidth
Werkzeug
yarl
zipp
